requirements.txt
absl-py==2.1.0
altair==5.4.1
appdirs==1.4.4
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dominate==2.8.0
filelock==3.16.1
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.6
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.12.1
html5lib==1.1
idna==3.10
imageio==2.36.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==4.9.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multitasking==0.0.11
mysql-connector-python==9.2.0
namex==0.0.8
narwhals==1.10.0
networkx==3.3
nltk==3.9.1
numpy==2.0.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==23.2
pandas==2.2.2
peewee==3.17.7
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.28.3
pyarrow==17.0.0
pyasynchat==1.0.4
pyasyncore==1.0.4
pydeck==0.9.1
pyftpdlib==2.0.1
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
pyshark==0.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-futures==1.0.1
rich==13.9.3
rpds-py==0.20.0
scapy==2.5.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==75.2.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.4.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vaderSentiment==3.3.2
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.44.0
wrapt==1.16.0
xgboost==2.1.2
yahooquery==2.3.7
yfinance==0.2.46
